basic bot that eats things

+++ src/aiPython/BotPlayer.js
@@ -1,7 +1,6 @@
 var zerorpc = require("zerorpc");
 var path = require('path');
 var child_process = require('child_process');
-var portfinder = require('portfinder');
 
 var PlayerTracker = require('../PlayerTracker');
 var gameServer = require('../GameServer');
@@ -10,6 +9,7 @@ function BotPlayer(currentGameServer, s, spawnPython, zerorpcPort) {
     PlayerTracker.apply(this, Array.prototype.slice.call(arguments));
     //this.color = gameServer.getRandomColor();
 
+    this.currentGameServer = currentGameServer;
     this.spawnPython = spawnPython;
     this.zerorpcPort = zerorpcPort
     // AI only
@@ -60,6 +60,16 @@ module.exports = BotPlayer;
 BotPlayer.prototype = new PlayerTracker();
 
 BotPlayer.prototype.update = function() { // Overrides the update function from player tracker
+    //if cell is dead 
+    if (this.cells[0] == null){
+        if (this.spawnPython) {
+            console.log('bot was eaten')
+            this.pythonLogic.kill();
+        }
+        this.socket.close();
+        return
+    }
+
     // Remove nodes from visible nodes if possible
     for (var i = 0; i < this.nodeDestroyQueue.length; i++) {
         var index = this.visibleNodes.indexOf(this.nodeDestroyQueue[i]);
@@ -80,17 +90,57 @@ BotPlayer.prototype.update = function() { // Overrides the update function from 
 
     // Calculate nodes
     this.visibleNodes = this.calcViewBox();
-    //console.log(this)
+
+    var that = this;
+    // var nodes = this.currentGameServer.nodes.map(function(node) {
+    //         if (that.cells[0].owner !== node.owner) {
+    //             return nodeSimp(node);
+    //         } else {
+    //             console.log('removing self')
+    //             return null;
+    //         }
+    //     }
+    // );
+    // var nodes = this.currentGameServer.nodes.reduce(function(previousValue, currentValue, currentIndex, array) {
+    //         if (that.cells[0].owner !== currentValue.owner) {
+    //             return previousValue.concat(nodeSimp(currentValue));
+    //         } else {
+    //             return previousValue;
+    //         }
+    // }, []);
+    // //console.log(this)
 
     gameInfo = {
-        a: 'b'
+        'cell': this.cells[0],
+        'nodes': this.visibleNodes.filter(function(node) {return node.owner !== that.cells[0].owner})
     }
 
-    this.zerorpcClient.invoke("getNewMousePosition", gameInfo, (function(error, res, more) {
+    var gameInfoString = JSON.stringify(gameInfo, function(key, value) {
+            if (typeof value === 'object' && value !== null) {
+                // console.log(value.owner)
+                try {
+                    if (!( value.constructor.name === "Cell"
+                        || value.constructor.name === "Object"
+                        || value.constructor.name === "Array")) {
+                    // Circular reference found, discard key
+                        return;
+                    }
+
+                } catch (err) {
+                    console.log(err)
+                    return 
+                }
+            }
+            return value;
+        }
+    );
+    // console.log(gameInfoString)
+    this.zerorpcClient.invoke("getNewMousePosition", gameInfoString, (function(error, res, more) {
             if (error) {
                 return console.log(error)
             }
-            //console.log(res.x + ", " + res.y)
+            // console.log(res.message)
+            // console.log(res.x + ", " + res.y)
             this.mouse = {
                 x: res.x,
                 y: res.y
@@ -98,13 +148,31 @@ BotPlayer.prototype.update = function() { // Overrides the update function from 
         }).bind(this)
     );
 
-
     // Reset queues
     this.nodeDestroyQueue = [];
     this.nodeAdditionQueue = [];
 };
 
+var nodeSimp = function(node) {
+    var newNode = {}
+    newNode.position = node.position;
+    newNode.mass = node.mass;
+    newNode.cellType = node.cellType;
+    newNode.size = node.size;
+    newNode.cellType = node.cellType;
+    newNode.angle = node.angle;
+    newNode.moveEngineSpeed = node.moveEngineSpeed;
+    return newNode;
+}
+
 BotPlayer.prototype.kill = function() {
     console.log('killing bot: ' + this.zerorpcPort)
-    this.pythonLogic.kill();
+    if (this.spawnPython) {
+        try {
+            this.pythonLogic.kill();
+        } catch (err) {
+            console.log(err)
+        }
+    }
+
 }
